pyproject.toml
[project]
name = "NapCatQQ-Desktop"
version = "1.7.28"
description = "为 NapCat 提供管理界面(GUI)，目的是让用户能够更快速、更直观的使用 NapCat"
authors = [{ name = "Qiao", email = "qiaohuayuquan@qq.com" }]
dependencies = [
    "httpx>=0.27.2",
    "pydantic>=2.9.0",
    "psutil>=6.0.0",
    "markdown>=3.7",
    "pyinstaller>=6.16.0",
    "PySide6-Fluent-Widgets[full]>=1.9.1",
    "creart>=0.3.0",
    "qrcode>=8.2",
    "pytest>=8.0.0",
    "pytest-qt>=4.4.0",
    "pytest-cov>=6.0.0",
]
requires-python = "==3.12.*"
readme = "README.md"
license = { text = "GPLv3" }

[tool.black]
line-length = 120          # 设置代码行的最大长度为 120 字符
target-version = ['py312'] # 目标 Python 版本为 3.12


[tool.isort]
known_first_party = ["NapCatQQ-Desktop"]
profile = "black"                        # 兼容 black
line_length = 120                        # 将行长度设置为与 black 一致

# 导入分类顺序：未来标准库、标准库、第三方库、PySide6 库、项目内库，和本地库
sections = [
    "FUTURE",
    "STDLIB",
    "THIRDPARTY",
    "PYSIDE6",
    "FIRSTPARTY",
    "LOCALFOLDER",
]
known_pyside6 = ["PySide6"] # 将 PySide6 标记为单独的分组
no_lines_before = ["PYSIDE6"] # 确保在 PYSIDE6 之前没有空行

# 导入自动注释
import_heading_stdlib = "标准库导入"
import_heading_thirdparty = "第三方库导入"
import_heading_firstparty = "项目内模块导入"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
markers = [
    "unit: 单元测试",
    "integration: 集成测试",
    "gui: GUI 测试",
    "slow: 运行较慢的测试",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/site-packages/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "if typing.TYPE_CHECKING:",
]
